Rename subjectRowDomId module to subjectRowAnchor
The module now holds both the row DOM-id builder/parser and the URL-fragment
parser (subjectIdFromHash), so naming it after only the DOM id undersells it —
and names the very concept its own comments call out as decoupled from the DOM
id. Rename the module and its spec to subjectRowAnchor, spanning both the
DOM-id and fragment forms of a row's anchor. Only import paths change.
